requirements.txt
# Core dependencies
numpy>=1.22.0
pandas>=1.4.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.25.0
diffusers>=0.12.0
accelerate>=0.15.0
safetensors>=0.3.0

# OCR
paddlepaddle>=2.4.0
paddleocr>=2.6.0

# Image processing
Pillow>=9.3.0
scikit-image>=0.19.0
opencv-python>=4.6.0

# Captioning
lavis>=0.3.0  # Contains BLIP-2 implementation

# Text processing
sentence-transformers>=2.2.2
nltk>=3.7

# Evaluation metrics
torchmetrics>=0.11.0
lpips>=0.1.4

# UI
streamlit>=1.16.0
streamlit-image-comparison>=0.0.3

# Utilities
pyyaml>=6.0
tqdm>=4.64.1
matplotlib>=3.6.0